SWEEP: Removed unnecessary casts of int or long to (sbyte) prior to cast to (byte), as overflow automatically fixes this

+++ src/Lucene.Net.Codecs/Memory/DirectDocValuesConsumer.cs
@@ -139,7 +139,7 @@ namespace Lucene.Net.Codecs.Memory
                 switch (byteWidth)
                 {
                     case 1:
-                        data.WriteByte((byte)(sbyte) v);
+                        data.WriteByte((byte) v);
                         break;
                     case 2:
                         data.WriteInt16((short) v);
